requirements.txt
boto==2.49.0
boto3==1.14.54
botocore==1.17.54
cachetools==4.1.1
certifi==2020.6.20
chardet==3.0.4
click==7.1.2
docutils==0.15.2
dominate==2.5.2
Flask==1.1.2
Flask-Bootstrap==3.3.7.1
Flask-Dropzone==1.5.4
gensim==3.8.3
google-api-core==1.22.1
google-auth==1.21.0
google-cloud-vision==1.0.0
googleapis-common-protos==1.52.0
grpcio==1.31.0
gunicorn==20.0.4
idna==2.10
itsdangerous==1.1.0
Jinja2==2.11.2
jmespath==0.10.0
joblib==0.16.0
MarkupSafe==1.1.1
nltk==3.5
numpy==1.19.1
pandas==1.1.1
protobuf==3.13.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
python-dateutil==2.8.1
pytz==2020.1
regex==2020.7.14
requests==2.24.0
rsa==4.6
s3transfer==0.3.3
scipy==1.5.2
six==1.15.0
smart-open==2.1.1
tqdm==4.48.2
urllib3==1.25.10
visitor==0.1.3
Werkzeug==1.0.1
